Updates - Modularity Clock & MemoryTest. - Change folder name from three to scenes. - Slightly simplify some codes of scenes.
