+++ src/components/ClockApp.module.scss
@@ -1,5 +1,4 @@
 $size: 90vmin;
-@import "../components/Section.module.scss";
 
 @font-face {
   font-family: "liquid_crystal-font";
