+++ src/components/Front.module.scss
