+++ src/components/RareBeef.module.scss
@@ -4,6 +4,7 @@
   height: 500vh;
   display: block;
 }
+
 .content {
   overflow: hidden;
   position: sticky;
