+++ src/components/ReactNative.module.scss
@@ -4,6 +4,7 @@
   height: 800vh;
   display: block;
 }
+
 .content {
   overflow: hidden;
   position: sticky;
