+++ src/routes/Home.tsx
@@ -1,16 +1,16 @@
 import React, { useRef } from "react";
-import ReactNative from "../sections/ReactNative";
-import Front from "../sections/Front";
 import styles from "./Home.module.scss";
 import classNames from "classnames";
 import { HomePropType } from "../types";
-import Clock from "../sections/Clock";
-import MemoryTest from "../sections/MemoryTest";
 import Section from "../components/Section";
+import Front from "../components/Front";
+import RareBeef from "../components/RareBeef";
 import placeReview from "../sections/placeReview";
+import ReactNative from "../components/ReactNative";
 import metaBeef from "../sections/metaBeef";
+import memoryTest from "../sections/memoryTest";
 import simpleMemo from "../sections/simpleMemo";
-import RareBeef from "../sections/RareBeef";
+import clock from "../sections/clock";
 import HomeHelmet from "../helmets/HomeHelmet";
 
 const Home: React.FC<HomePropType> = () => {
@@ -25,9 +25,9 @@ const Home: React.FC<HomePropType> = () => {
       <Section data={placeReview} />
       <ReactNative />
       <Section data={metaBeef} />
-      <MemoryTest />
+      <Section data={memoryTest} />
       <Section data={simpleMemo} />
-      <Clock />
+      <Section data={clock} />
       <footer className={styles.footer}>
         &copy; {new Date().getFullYear()}. RAREBEEF All Rights Reserved.
       </footer>
