+++ src/scenes/Phones.tsx
@@ -30,33 +30,28 @@ const Phones: React.FC<PhonesPropType> = ({ sectionRef }) => {
 
   useEffect(() => {
     const windowScrollListner = () => {
-      if (
-        !sectionRef.current ||
-        sectionRef.current.getBoundingClientRect().top /
-          (window.innerHeight * 8 -
-            sectionRef.current.childNodes[1].clientHeight) >=
-          0 ||
-        sectionRef.current.getBoundingClientRect().top /
-          (window.innerHeight * 8 -
-            sectionRef.current.childNodes[1].clientHeight) <=
-          -1.5
-      ) {
+      if (!sectionRef.current || !controlRef.current) {
         return;
       }
 
+      const controlPos = controlRef.current.object.position;
+      const setcion = sectionRef.current;
       // (뷰포트 상단 기준 section top의 y 위치) / (800vh - sticy요소(content) 높이)
       // fixed 시작 시점이 0,
       // fixed가 해제될 때 100이 된다.
       let scrollDegree =
-        (sectionRef.current.getBoundingClientRect().top /
-          (window.innerHeight * 8 -
-            sectionRef.current.childNodes[1].clientHeight)) *
+        (setcion.getBoundingClientRect().top /
+          (window.innerHeight * 8 - setcion.childNodes[1].clientHeight)) *
         -1;
+
+      if (scrollDegree <= 0 || scrollDegree >= 1.5) {
+        return;
+      }
 
       if (scrollDegree >= 0 && scrollDegree < 0.2) {
         scrollDegree *= 5;
 
-        gsap.to(controlRef.current.object.position, 0.3, {
+        gsap.to(controlPos, 0.3, {
           x: -5 + 5 * scrollDegree,
           y: -1,
           z: 0 + 3 * scrollDegree,
@@ -65,7 +60,7 @@ const Phones: React.FC<PhonesPropType> = ({ sectionRef }) => {
       } else if (scrollDegree >= 0.2 && scrollDegree < 0.4) {
         scrollDegree = (scrollDegree - 0.2) * 5;
 
-        gsap.to(controlRef.current.object.position, 0.3, {
+        gsap.to(controlPos, 0.3, {
           x: 0 + 5 * scrollDegree,
           y: -1,
           z: 3 - 3 * scrollDegree,
@@ -74,7 +69,7 @@ const Phones: React.FC<PhonesPropType> = ({ sectionRef }) => {
       } else if (scrollDegree >= 0.4 && scrollDegree < 0.6) {
         scrollDegree = (scrollDegree - 0.4) * 5;
 
-        gsap.to(controlRef.current.object.position, 0.3, {
+        gsap.to(controlPos, 0.3, {
           x: 5 - 5 * scrollDegree,
           y: -1,
           z: 0 - 1 * scrollDegree,
@@ -83,7 +78,7 @@ const Phones: React.FC<PhonesPropType> = ({ sectionRef }) => {
       } else if (scrollDegree >= 0.6 && scrollDegree < 0.8) {
         scrollDegree = (scrollDegree - 0.6) * 5;
 
-        gsap.to(controlRef.current.object.position, 0.3, {
+        gsap.to(controlPos, 0.3, {
           x: 0,
           y: -1 - 5.5 * scrollDegree,
           z: -1,
@@ -92,7 +87,7 @@ const Phones: React.FC<PhonesPropType> = ({ sectionRef }) => {
       } else if (scrollDegree >= 0.8 && scrollDegree <= 1.05) {
         scrollDegree = (scrollDegree - 0.8) * 5;
 
-        gsap.to(controlRef.current.object.position, 0.3, {
+        gsap.to(controlPos, 0.3, {
           x: 0 - 5 * scrollDegree,
           y: -6.5 + 6 * scrollDegree,
           z: -1 + 6 * scrollDegree,
