--- src/sections/MemoryTest.module.scss
@@ -1,2 +1,0 @@
-@import "../components/Section.module.scss";
-// 앱 상하 여백은 components/MemoryTestApp.module.scss 에서 수정
